+++ outputs/elasticsearch/output.go
@@ -67,7 +67,7 @@ func (out *elasticsearchOutput) Init(
 	if len(config.Hosts) > 0 {
 		// use hosts setting
 		for _, host := range config.Hosts {
-			url, err := getUrl(host)
+			url, err := getUrl(config.Protocol, config.Path, host)
 
 			if err != nil {
 				logp.Err("Invalid host param set: %s, Error: %v", host, err)
@@ -75,8 +75,7 @@ func (out *elasticsearchOutput) Init(
 			urls = append(urls, url)
 		}
 	} else {
-		// use host and port settings
-		// TODO: Deprecate usage of host, always use hosts
+		// usage of host and port is deprecated as it is replaced by hosts
 		url := fmt.Sprintf("%s://%s:%d%s", config.Protocol, config.Host, config.Port, config.Path)
 		urls = append(urls, url)
 	}
@@ -293,7 +292,7 @@ func (out *elasticsearchOutput) BulkPublish(
 
 // Creates the url based on the url configuration.
 // Adds missing parts with defaults (scheme, host, port)
-func getUrl(rawUrl string) (string, error) {
+func getUrl(defaultScheme string, defaultPath string, rawUrl string) (string, error) {
 
 	urlStruct, err := url.Parse(rawUrl)
 
@@ -304,7 +303,7 @@ func getUrl(rawUrl string) (string, error) {
 	host := ""
 	port := ""
 
-	// If url doesn't have a scheme, host is written into path
+	// If url doesn't have a scheme, host is written into path. For example: 192.168.3.7
 	if urlStruct.Host == "" {
 		urlStruct.Host = urlStruct.Path
 		urlStruct.Path = ""
@@ -335,7 +334,12 @@ func getUrl(rawUrl string) (string, error) {
 
 	// Assign default scheme if not set
 	if urlStruct.Scheme == "" {
-		urlStruct.Scheme = "http"
+		urlStruct.Scheme = defaultScheme
+	}
+
+	// Assign default path if not set
+	if urlStruct.Path == "" {
+		urlStruct.Path = defaultPath
 	}
 
 	// Check if ipv6