+++ CHANGELOG.md
@@ -15,3 +15,6 @@ All notable changes to this project will be documented in this file based on the
 - Add tls configuration support to elasticsearch and lumberjack outputers
 
 ### Deprecated
+
+ * host and port configuration options are deprecated. They are replaced by the hosts
+ configuration option.